+++ micropip/_compat_not_in_pyodide.py
@@ -64,45 +64,6 @@ def canonicalize_name(name: str) -> str:
     return _canonicalize_regex.sub("-", name).lower()
 
 
-# Vendored from pip
-class UnsupportedWheel(Exception):
-    """Unsupported wheel."""
-
-
-def wheel_dist_info_dir(source: ZipFile, name: str) -> str:
-    """Returns the name of the contained .dist-info directory.
-    Raises UnsupportedWheel if not found, >1 found, or it doesn't match the
-    provided name.
-    """
-    # Zip file path separators must be /
-    subdirs = {p.split("/", 1)[0] for p in source.namelist()}
-
-    info_dirs = [s for s in subdirs if s.endswith(".dist-info")]
-
-    if not info_dirs:
-        raise UnsupportedWheel(f".dist-info directory not found in wheel {name!r}")
-
-    if len(info_dirs) > 1:
-        raise UnsupportedWheel(
-            "multiple .dist-info directories found in wheel {!r}: {}".format(
-                name, ", ".join(info_dirs)
-            )
-        )
-
-    info_dir = info_dirs[0]
-
-    info_dir_name = canonicalize_name(info_dir)
-    canonical_name = canonicalize_name(name)
-    if not info_dir_name.startswith(canonical_name):
-        raise UnsupportedWheel(
-            ".dist-info directory {!r} does not start with {!r}".format(
-                info_dir, canonical_name
-            )
-        )
-
-    return info_dir
-
-
 class pyodide_js_:
     def __get__(self, attr):
         raise RuntimeError(f"Attempted to access property '{attr}' on pyodide_js dummy")
@@ -124,6 +85,5 @@ __all__ = [
     "loadedPackages",
     "loadPackage",
     "get_dynlibs",
-    "wheel_dist_info_dir",
     "to_js",
 ]
