+++ micropip/metadata.py
@@ -3,8 +3,19 @@ This is a stripped down version of pip._vendor.pkg_resources.DistInfoDistributio
 """
 from pathlib import Path
 from packaging.requirements import Requirement
+from packaging.utils import canonicalize_name
 import re
-from collections.abc import Sequence 
+from collections.abc import Sequence
+from zipfile import ZipFile
+
+
+def safe_name(name):
+    """Convert an arbitrary string to a standard distribution name
+
+    Any runs of non-alphanumeric/. characters are replaced with a single '-'.
+    """
+    return re.sub("[^A-Za-z0-9.]+", "-", name)
+
 
 def safe_extra(extra):
     """Convert an arbitrary string to a standard 'extra' name
@@ -14,6 +25,44 @@ def safe_extra(extra):
     """
     return re.sub("[^A-Za-z0-9.-]+", "_", extra).lower()
 
+
+# Vendored from pip
+class UnsupportedWheel(Exception):
+    """Unsupported wheel."""
+
+
+def wheel_dist_info_dir(source: ZipFile, name: str) -> str:
+    """Returns the name of the contained .dist-info directory.
+    Raises UnsupportedWheel if not found, >1 found, or it doesn't match the
+    provided name.
+    """
+    # Zip file path separators must be /
+    subdirs = {p.split("/", 1)[0] for p in source.namelist()}
+
+    info_dirs = [s for s in subdirs if s.endswith(".dist-info")]
+
+    if not info_dirs:
+        raise UnsupportedWheel(f".dist-info directory not found in wheel {name!r}")
+
+    if len(info_dirs) > 1:
+        raise UnsupportedWheel(
+            "multiple .dist-info directories found in wheel {!r}: {}".format(
+                name, ", ".join(info_dirs)
+            )
+        )
+
+    info_dir = info_dirs[0]
+
+    info_dir_name = canonicalize_name(info_dir)
+    canonical_name = canonicalize_name(name)
+    if not info_dir_name.startswith(canonical_name):
+        raise UnsupportedWheel(
+            ".dist-info directory {!r} does not start with {!r}".format(
+                info_dir, canonical_name
+            )
+        )
+
+    return info_dir
 
 class Metadata:
     """
